fix(audit): К1 третья партия — маски текстур 64-битно, протухшие id prop.frag, решётка населения
По карте core-stabilization.md:

1. МАСКИ КАРТ МАТЕРИАЛОВ (К1-2): id рыхлых двойников доходят до 37 —
   прежний путь был сломан ТРИЖДЫ: UB `1u<<32` при загрузке (rubble_rust),
   ВЕЛИЧИННЫЙ каст float(torus.z) корёжил округлением всю маску альбедо
   (мантисса 24 бита < старших битов), упаковка 16+16 в torus.w молча
   отрезала normal/roughness у id >= 16. Теперь маски uint64 по построению
   и едут парами u32 в MarchUbo (1136 Б, static_assert); пуш-лейны
   torus.z/w свободны — двойное назначение лейны (ловушка C4) умерло.
   Лог печатает маску полностью (%llx: 0x1f000fc00, 33-й бит на месте).
2. prop.frag (К1-3): анизотропный блик «для труб» держался за id 3/4
   СТАРОЙ таблицы — после роста до 38 строк светил на ВОДЕ и плите —
   вырезан (вернётся колонкой CSV по вкусу); порог `mat_id >= 3u` в
   compute_prop_roughness перехватывал все материалы — три семейства
   шершавости (ribbed/rust+rubble/plaster+plank) ожили.
3. РЕШЁТКА НАСЕЛЕНИЯ (К1-4): смещение 12 кратно страйду 4 таксономии
   комнат — каждый четвёртый житель спавнился «в стене» (room_bit 0) и
   его errand молча не работал. {3,6,9,12} -> {3,6,9,13}.

ctest 7/7, ноль варнингов.

+++ src/render/cube_pass.h
@@ -78,9 +78,13 @@ public:
 
     // Bit i set == material id i draws from a real photograph rather than the
     // procedural surface. 0 means the whole pack fell back.
-    std::uint32_t textured_materials() const { return texMask_; }
-    std::uint32_t normal_materials() const { return normalMask_; }
-    std::uint32_t roughness_materials() const { return roughnessMask_; }
+    // Маски «у материала есть карта» — uint64 ПО ПОСТРОЕНИЮ (аудит
+    // 2026-08-25, К1-2): id двойников доходят до 37, прежний u32 давал UB
+    // `1u<<32`, а упаковка в float(torus.z) с ростом id за 24 бита мантиссы
+    // корёжила ВСЮ маску округлением. Возятся в MarchUbo, не в пуше.
+    std::uint64_t textured_materials() const { return texMask_; }
+    std::uint64_t normal_materials() const { return normalMask_; }
+    std::uint64_t roughness_materials() const { return roughnessMask_; }
 
     // Borrowed by passes that share the same push-constant block (PropPass).
     VkPipelineLayout pipeline_layout() const { return layout_; }
@@ -102,9 +106,9 @@ private:
     VulkanTextureArray normal_;
     VulkanTextureArray roughness_;
 
-    std::uint32_t texMask_ = 0;
-    std::uint32_t normalMask_ = 0;
-    std::uint32_t roughnessMask_ = 0;
+    std::uint64_t texMask_ = 0;
+    std::uint64_t normalMask_ = 0;
+    std::uint64_t roughnessMask_ = 0;
     bool textured_ = false;
 };
